+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/DebugColorizeTilesRasterOverlay.h
@@ -22,16 +22,15 @@ public:
   /**
    * @copydoc RasterOverlay::createTileProvider
    */
-  virtual CesiumAsync::Future<
-      CesiumUtility::IntrusivePointer<RasterOverlayTileProvider>>
-  createTileProvider(
+  virtual CesiumAsync::Future<CreateTileProviderResult> createTileProvider(
       const CesiumAsync::AsyncSystem& asyncSystem,
       const std::shared_ptr<CesiumAsync::IAssetAccessor>& pAssetAccessor,
       const std::shared_ptr<CreditSystem>& pCreditSystem,
       const std::shared_ptr<IPrepareRendererResources>&
           pPrepareRendererResources,
       const std::shared_ptr<spdlog::logger>& pLogger,
-      const RasterOverlay* pOwner) const override;
+      CesiumUtility::IntrusivePointer<const RasterOverlay> pOwner)
+      const override;
 };
 
 } // namespace Cesium3DTilesSelection