+++ Cesium3DTilesSelection/src/Tile.cpp
@@ -181,7 +181,13 @@ int64_t Tile::computeByteSize() const noexcept {
 }
 
 bool Tile::isRenderable() const noexcept {
-  if (getState() >= TileLoadState::ContentLoaded) {
+  if (getState() == TileLoadState::Failed) {
+    // Explicitly treat failed tiles as "renderable" - we just treat them like
+    // empty tiles.
+    return true;
+  }
+
+  if (getState() == TileLoadState::Done) {
     if (!isExternalContent()) {
       return std::all_of(
           this->_rasterTiles.begin(),