+++ Cesium3DTilesSelection/src/TilesetContentManager.h
@@ -43,11 +43,21 @@ public:
       const std::string& ionAccessToken,
       const std::string& ionAssetEndpointUrl = "https://api.cesium.com/");
 
+  /**
+   * @brief A future that resolves after all async operations initiated by this
+   * content manager have completed and all tiles are unloaded, but before the
+   * content manager itself is destroyed.
+   */
+  CesiumAsync::SharedFuture<void>& getAsyncDestructionCompleteEvent();
+
   ~TilesetContentManager() noexcept;
 
   void loadTileContent(Tile& tile, const TilesetOptions& tilesetOptions);
 
-  void updateTileContent(Tile& tile, const TilesetOptions& tilesetOptions);
+  void updateTileContent(
+      Tile& tile,
+      double priority,
+      const TilesetOptions& tilesetOptions);
 
   bool unloadTileContent(Tile& tile);
 
@@ -87,16 +97,24 @@ public:
 
   bool tileNeedsLoading(const Tile& tile) const noexcept;
 
+  void tickMainThreadLoading(
+      double timeBudget,
+      const TilesetOptions& tilesetOptions);
+
 private:
   static void setTileContent(
       Tile& tile,
       TileLoadResult&& result,
       void* pWorkerRenderResources);
 
-  void
-  updateContentLoadedState(Tile& tile, const TilesetOptions& tilesetOptions);
+  void updateContentLoadedState(
+      Tile& tile,
+      double priority,
+      const TilesetOptions& tilesetOptions);
 
   void updateDoneState(Tile& tile, const TilesetOptions& tilesetOptions);
+
+  void finishLoading(Tile& tile, const TilesetOptions& tilesetOptions);
 
   void unloadContentLoadedState(Tile& tile);
 
@@ -126,5 +144,25 @@ private:
   int32_t _tilesLoadOnProgress;
   int32_t _loadedTilesCount;
   int64_t _tilesDataUsed;
+
+  struct MainThreadLoadTask {
+    Tile* pTile;
+
+    /**
+     * @brief The relative priority of loading for this tile.
+     *
+     * Lower priority values load sooner.
+     */
+    double priority;
+
+    bool operator<(const MainThreadLoadTask& rhs) const noexcept {
+      return this->priority < rhs.priority;
+    }
+  };
+
+  std::vector<MainThreadLoadTask> _finishLoadingQueue;
+
+  CesiumAsync::Promise<void> _destructionCompletePromise;
+  CesiumAsync::SharedFuture<void> _destructionCompleteFuture;
 };
 } // namespace Cesium3DTilesSelection